refactor: relocate extension update-check + is_git_url utils to core
Extension-crate extraction prep (1 of ~4 remaining edges). check_update_available
(+ UpdateAvailable) and is_git_url are core glue misfiled in the extension
lifecycle module: check_update_available inspects an extension's git checkout via
paths/git/is_extension_linked (no extension behavior), and is_git_url is a pure
URL-prefix string check.

Relocates both to crate::extension_update_check. The extension module re-exports
them for path stability (crate::extension::{is_git_url,check_update_available,
UpdateAvailable} still resolve), so consumers - context/report.rs (core),
homeboy-upgrade, homeboy-rig - stay stable.

Breaks two of the last core->extension edges surfaced when attempting the
wholesale homeboy-extension git-mv.
