+++ crates/homeboy-core/src/extension/lifecycle/update.rs
@@ -399,63 +399,6 @@ fn linked_extension_git_root(extension_id: &str, source_dir: &Path) -> Result<Pa
     )))
 }
 
-/// Check if a git-cloned extension has updates available.
-/// Runs `git fetch` then checks if HEAD is behind the remote tracking branch.
-/// Returns None for linked extensions or if check fails.
-pub fn check_update_available(extension_id: &str) -> Option<UpdateAvailable> {
-    let extension_dir = paths::extension(extension_id).ok()?;
-    if !extension_dir.exists() || is_extension_linked(extension_id) {
-        return None;
-    }
-
-    // Check it's a git repo
-    if !extension_dir.join(".git").exists() {
-        return None;
-    }
-
-    // Fetch latest (best-effort, short timeout)
-    Command::new("git")
-        .args(["fetch", "--quiet"])
-        .current_dir(&extension_dir)
-        .stdin(std::process::Stdio::null())
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .status()
-        .ok()?;
-
-    // Check how many commits we're behind
-    let output = Command::new("git")
-        .args(["rev-list", "HEAD..@{u}", "--count"])
-        .current_dir(&extension_dir)
-        .stdin(std::process::Stdio::null())
-        .output()
-        .ok()?;
-
-    let count_str = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    let behind_count: usize = count_str.parse().ok()?;
-
-    if behind_count == 0 {
-        return None;
-    }
-
-    // Get installed version
-    let extension = load_extension(extension_id).ok()?;
-    let installed_version = extension.version.clone();
-
-    Some(UpdateAvailable {
-        extension_id: extension_id.to_string(),
-        installed_version,
-        behind_count,
-    })
-}
-
-#[derive(Debug, Clone)]
-pub struct UpdateAvailable {
-    pub extension_id: String,
-    pub installed_version: String,
-    pub behind_count: usize,
-}
-
 /// Read the source revision for an installed extension.
 /// Checks (in order): .git directory (git rev-parse), then .source-revision file.
 pub fn read_source_revision(extension_id: &str) -> Option<String> {
