+++ crates/homeboy-core/src/lib.rs
@@ -101,6 +101,7 @@ pub mod extension_invocation_context;
 pub mod extension_provider_discovery;
 pub mod extension_scope;
 pub mod extension_store;
+pub mod extension_update_check;
 // finding moved to the internal `homeboy-finding` crate. Re-exported so existing
 // `crate::finding::*` call sites keep working unchanged.
 pub use homeboy_finding as finding;
